Breakup timeseries indicator sub-modules
